Add filterable concern to mushroom class
The Filterable concern filter method takes filter_attributes (array of strings) and
params. It iterates over the passed enum attributes and queries each of them by name
while holding the params[attribute] as value. It only return results if the params[attribute]
is present. Thus returning only the params we ask for which is nifty.