[WIP] Create a bunch of rake import helper methods
Add import helpers module that has a method for each attribute.
These method will take a string letter from a line and return
an attribute value if a line starts with 'p' then

calling ImportHelper.edible(line[0]) will return 'poisonous'

It basically queries the hash for is key and turns that a string.

{ edible: 'e', poisonous: 'p' }.key(arg).to_s

This approche is a bit repetitive but it's fine for now.
Rake tasks are soposed to be 1 offs. and all our attributes are known.
I dont think the dataset structure will change anytime soon. so its fine.