Set database production url logic and update schema
When the app is inside a docker it and its on heroku
it should use the DATABAS_URL variable or grab it from the cre.

When running docker locally with docker compose use db instead of localhost.

when running rails in production mode without docker arround it use default
localhost and mushrooms_production database.

U