Add docker-compose file for local usage
again here we also set the DOCKENV variable to true for the database.yml logic.
docker-compose reads the .env file which hold the RAILS_MASTER_KEY=... with this:

environment:
  RAILS_MASTER_KEY=$RAILS_MASTER_KEY

my app depends on postgres and it pull the container image of the app it self
from dockerhub i allready build it and pushed it there.