Add bin/docker-run command
after running the `bin/docker-setup` command, lets say we quite
and launch the app later on, we dont want to run the setup again.

so we just do `bin/docker-run`

what this basically does is a docker-compose up but i like to keep
things here maybe there will be more commands todo in the future
like runnig a worker or whatever... putting them all in one place
make life easy.

now a little front end stuff and we can really ship it.